code/requirements.txt
torch
torchvision
diffusers
opencv-python
Pillow
numpy
lpips